+++ module2/service/src/main/java/com/epam/esm/service/validator/GiftCertificateValidator.java
@@ -1,15 +1,17 @@
 package com.epam.esm.service.validator;
 
+import java.math.BigDecimal;
+
 public interface GiftCertificateValidator {
     boolean isNameValid(String name);
 
     boolean isDescriptionValid(String description);
 
-    boolean isPriceValid(double price);
+    boolean isPriceValid(BigDecimal price);
 
     boolean isDurationValid(short duration);
 
-    boolean validateAll(String name, String description, double price, short duration);
+    boolean validateAll(String name, String description, BigDecimal price, short duration);
 
     boolean isSortTypeValid(String sortType);
